Cleanup and consolidate cursor test code

+++ src/main/java/com/facebook/presto/block/rle/RunLengthEncodedCursor.java
@@ -126,7 +126,7 @@ public class RunLengthEncodedCursor
     {
         Preconditions.checkState(block != null, "Need to call advanceNext() first");
 
-        return block.getSingleValue().getLong(field);
+        return block.getSingleValue().getDouble(field);
     }
 
     @Override
